Rename layout.py to _layout_utils.py
